test: type bulk-memory address as usize in TestWASMBulkMemoryPrimitives (F2)
F2 gates the implicit i32<->usize conversion to stdlib context, so this
user-level test program calling __memset / __memcpy with an i32 address
no longer type-checks. Type the slice data pointer as usize (the correct
type for a memory address) via an explicit `as usize` cast; arithmetic on
it stays usize and flows into the usize params with no per-call cast. On
wasm32 usize is 4 bytes (== i32) so codegen and runtime are unchanged —
only the type-check differs. Caught by the wasm e2e shard (no local wasm
toolchain to run it pre-push).

+++ internal/e2e/wasm_e2e_test.go
@@ -2608,7 +2608,11 @@ func TestWASMBulkMemoryPrimitives(t *testing.T) {
         77 as u8, 78 as u8, 79 as u8, 80 as u8
     ];
     var bs: [u8] = buf[0:16];
-    var base: i32 = (bs as i32);
+    // The data pointer is a memory address — type it usize (pointer
+    // width) so it flows into the usize-typed __memset / __memcpy params.
+    // Under F2, user code must reach usize via an explicit cast rather
+    // than the implicit i32<->usize hop. See docs/ADVERSARIAL-REVIEW-2026-06.md.
+    var base: usize = (bs as usize);
     __memset(base + 8, 0, 8);
     if (buf[0]  != 65)  { return 1; }
     if (buf[7]  != 72)  { return 2; }
